Rename module from dispatcher to invoker.
