+++ src/net/server/dgram.rs
@@ -46,7 +46,7 @@ use crate::net::server::util::to_pcap_text;
 use crate::utils::config::DefMinMax;
 
 use super::buf::VecBufSource;
-use super::dispatcher::{InvokerStatus, ServiceInvoker};
+use super::invoker::{InvokerStatus, ServiceInvoker};
 use super::message::{TransportSpecificContext, UdpTransportContext};
 use super::ServerCommand;
