+++ src/net/server/mod.rs
@@ -182,7 +182,7 @@ pub use connection::Config as ConnectionConfig;
 
 pub mod buf;
 pub mod dgram;
-pub mod dispatcher;
+pub mod invoker;
 pub mod error;
 pub mod message;
 pub mod metrics;
